requirements.txt
flask==3.0.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
werkzeug==3.0.1
gunicorn==21.2.0